rrhh: auditoría seguridad/UX + branding + cambiador empresa + ubicación fichajes (#941)
* fix(rrhh): auditoría completa jul-2026 — 4 críticos seguridad + 6 bugs código + UX

Seguridad:
- SEG-01: eliminado bypass User-Agent en los 3 crons; ahora solo acepta Bearer CRON_SECRET
  (sin fallback || '' que aceptaba token vacío)
- SEG-02: listarExpediente añade AND empresa_id al SELECT (antes filtraba solo por empleado_id)
- SEG-03: borrarDocumento invierte orden (Storage antes que BD) y DELETE scoped por empresa_id
- SEG-04: radio_m::float en SELECT de obras (era NUMERIC→string→NaN en dentroDeGeocerca)

Bugs:
- BUG-01: Fragment con key explícita en tabla de fichajes (evita crash reconciliación React)
- BUG-02: comparación timestamp por getTime() en lugar de comparación de string
- BUG-03: Promise con .catch() en carga de empleados (antes unhandled rejection)
- BUG-04: alert() nativo reemplazado por setErr/setAltaErr en EmpleadosClient y ObrasClient
- BUG-05: skeleton de carga en FichajeEmpleado (antes return null → pantalla en blanco)
- BUG-06: botones de acción py-1.5 min-h-[36px] en los 3 clientes admin (antes ~22px táctil)

Docs: informe completo en docs/AUDITORIA-2026-07-rrhh.md


* feat(rrhh): switch de empresa en sidebar + login neutro + branding Mariscos González

- Login muestra 'ia·rrhh' genérico (ya no carga la primera empresa aleatoria de BD)
- Nuevo endpoint GET /api/admin/mis-empresas — lista las empresas del usuario autenticado
- Nuevo endpoint POST /api/auth/cambiar-empresa — emite nuevo JWT sin cerrar sesión
- Nuevo componente CambiadorEmpresa: se auto-carga; si el responsable tiene >1 empresa
  muestra un selector al fondo del sidebar con logos miniatura y cambio sin salir
- BD: color_primario de Mariscos González actualizado a #1B3461 (azul marino corporativo)


* rrhh: mostrar ubicación GPS en fichajes sin obra asignada

Cuando un fichaje tiene coordenadas pero no coincide con ninguna
geovalla, la columna Obra muestra un enlace 📍 Ver mapa a Google Maps
con las coordenadas de entrada. Antes aparecía siempre '—'.


---------

+++ apps/rrhh/lib/fichajes.ts
@@ -18,7 +18,7 @@ export type Fichaje = {
 async function detectarObra(empresaId: string, lat: number | null, lng: number | null): Promise<string | null> {
   if (lat == null || lng == null) return null
   const obras = await prisma.$queryRaw<{ id: string; lat: number; lng: number; radio_m: number }[]>(Prisma.sql`
-    SELECT id, lat::float, lng::float, radio_m FROM rrhh.obras
+    SELECT id, lat::float, lng::float, radio_m::float FROM rrhh.obras
     WHERE empresa_id = ${empresaId}::uuid AND activa = true AND lat IS NOT NULL AND lng IS NOT NULL`)
   for (const o of obras) {
     if (dentroDeGeocerca({ lat, lng }, { lat: o.lat, lng: o.lng }, o.radio_m)) return o.id
